DeFi/YaoShuai-hub/liquidation-monitor/pyproject.toml
[project]
name = "liquidation-monitor"
version = "0.1.0"
description = "Morpho Blue Liquidation Monitor for HyperEVM - PKUBA Colearn"
requires-python = ">=3.11"
dependencies = [
    "web3>=7.0.0",
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0.0",
    "asyncio-throttle>=1.0.0",
    "websockets>=13.0",
]

[project.scripts]
monitor = "src.monitor:main"
monitor-all = "src.monitor_all:main"
digest = "src.cli:main"
backtest = "src.backtest:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
